Move FullTextSearch completely into it's own module instead of integrating into EntityInfo
- cleaner separation, especially that fulltext is optional makes this better
- no real problem to have an additional view
- no special handling to exclude duplicate PageTree Documents is needed
