+++ packages/api/cms-api/src/entity-info/entity-info.object.ts
@@ -1,5 +1,4 @@
 import { Entity, PrimaryKey, Property } from "@mikro-orm/core";
-import { FullTextType } from "@mikro-orm/postgresql";
 import { Field, ObjectType } from "@nestjs/graphql";
 
 // Note: This file is intentionally not named *.entity.ts to exclude it from MikroORM's CLI migration glob pattern.
@@ -26,7 +25,4 @@ export class EntityInfoObject {
 
     @Property({ type: "boolean" })
     visible: boolean;
-
-    @Property({ type: FullTextType })
-    fullText: string;
 }
