--- packages/api/cms-api/src/entity-info/full-text-search.module.ts
@@ -1,11 +1,0 @@
-import { Module } from "@nestjs/common";
-
-import { EntityInfoModule } from "./entity-info.module";
-import { FullTextSearchResolver } from "./full-text-search.resolver";
-
-/** @experimental */
-@Module({
-    imports: [EntityInfoModule],
-    providers: [FullTextSearchResolver],
-})
-export class FullTextSearchModule {}
