+++ packages/api/cms-api/src/full-text-search/dto/paginated-entity-info.ts
@@ -1,7 +1,7 @@
 import { ObjectType } from "@nestjs/graphql";
 
 import { PaginatedResponseFactory } from "../../common/pagination/paginated-response.factory";
-import { EntityInfoObject } from "../entity-info.object";
+import { EntityInfoObject } from "../../entity-info/entity-info.object";
 
 @ObjectType()
 export class PaginatedEntityInfo extends PaginatedResponseFactory.create(EntityInfoObject) {}
